+++ client/error.vue
@@ -1,3 +1,0 @@
-<template>
-  <NuxtLayout> Some 404 page here idk </NuxtLayout>
-</template>