+++ client/pages/about/index.vue
@@ -1,5 +1,0 @@
-<template>
-  <div class="prose-headings:font-heading">
-    <ContentDoc />
-  </div>
-</template>